Changed "unsigned long" to size_t in the prototypes of mmsafe allocators
Changed the size parameter's type from "unsigned long" to size_t
in all the mmsafe versions of heap allocators, because the original
malloc & realloc use size_t.

+++ include/safe_mm_checked.h
@@ -6,13 +6,14 @@
 #ifndef _SAFE_MM_CHECKED_H
 #define _SAFE_MM_CHECKED_H
 
+#include <stdlib.h>
 #include "stdchecked.h"
 
-for_any(T) mm_ptr<T> mm_alloc(unsigned long size);
+for_any(T) mm_ptr<T> mm_alloc(size_t size);
 for_any(T) void mm_free(mm_ptr<T> p);
 
-for_any(T) mm_array_ptr<T> mm_array_alloc(unsigned long array_size);
-for_any(T) mm_array_ptr<T> mm_array_realloc(mm_array_ptr<T> p, unsigned long size);
+for_any(T) mm_array_ptr<T> mm_array_alloc(size_t array_size);
+for_any(T) mm_array_ptr<T> mm_array_realloc(mm_array_ptr<T> p, size_t size);
 for_any(T) void mm_array_free(mm_array_ptr<T> p);
